55: upgrade the SDK to support Laravel 5.6 logging best practices
